requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.24
python-multipart>=0.0.6
httpx>=0.27.0
python-dotenv>=1.0.1
black>=24.0.0
ruff>=0.5.0
isort>=5.13.0
mypy>=1.10.0
